Adding old BST assignment
Copied in my solution to a BST problem from last year. Think some of the functions are unnecessary, but will work with the bits I need. Next step write a test for LCA function.

Bit of an isue with having to changee all BST occurences to LCA, so the naming seems off.