Added findLCA method
This works by first making sure the graph is a DAG, then reversing the dag. It uses BFS to find the two parameter values v and w (returns -1 if they are not found), and then cycles through both bfs paths (arr1 and arr2) and adds all the common points to a new array (two vertices can have more than one lca) and it then returns the first array (first one found, but it is reversed so it is the vertice furthest along the graph).